Refactor rating functions to live in shared helper rating overview object
