+++ client/src/components/overview/Rating.jsx
@@ -4,7 +4,7 @@ import StarRating from '../shared/StarRating.jsx';
 const Rating = ({ rating }) => {
   return (
     <div>
-      <StarRating score={rating.value} />
+      <StarRating score={rating.roundedValue} />
       <a href="">Read all {rating.numReviews} reviews</a>
     </div>
   );
